requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
Pillow==10.3.0
numpy==1.26.4

# TFLite — choose ONE of these (uncomment the right one):
# Option A: Lightweight (recommended — smaller install, faster)
tflite-runtime==2.14.0

# Option B: Full TensorFlow (larger but more compatible)
# tensorflow==2.14.0

# Testing
requests==2.32.3